Algorihtm_py changes
I checked the readme.md file from Spatieno which says that 'ImageTestResult' == 0 says no data (but we used -1 as no data, which is actually broken data). I therefore included the correction of this to Algorithm.py.

I have included the -1 anyways, if it says that the result is 'ok' but no data (which I don't think will actually happen). We should look at the 'testResult' for the broken data - also why I included a TODO.